+++ brightbox/data_source_brightbox_server_group.go
@@ -28,6 +28,16 @@ func dataSourceBrightboxServerGroup() *schema.Resource {
 				Optional:    true,
 				Computed:    true,
 			},
+			"default": {
+				Description: "Is this the default group for the account?",
+				Type:        schema.TypeBool,
+				Computed:    true,
+			},
+			"fqdn": {
+				Description: "Fully Qualified Domain Name",
+				Type:        schema.TypeString,
+				Computed:    true,
+			},
 		},
 	}
 }